requirements.txt
djangorestframework
channels
stripe
googlemaps
django-mailgun